+++ adbyss_psl/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 
 
+## [0.18.0](https://github.com/Blobfolio/adbyss/releases/tag/v0.18.0) - 2025-06-01
+
+### Changed
+
+* Bump `trimothy` to `0.8`
+* `Domain::has_www` is now const
+* `Domain::subdomain` is now const
+* `Domain::suffix` is now const
+* `Domain::tld` is now const
+* Lighten `Domain` struct by 16 bytes
+* Miscellaneous code cleanup and lints
+* Update suffix database
+
+### Breaking
+
+* Removed `Default` for `Domain`
+* Removed `Deref` for `Domain`
+
+
+
 ## [0.17.0](https://github.com/Blobfolio/adbyss/releases/tag/v0.17.0) - 2025-05-15
 
 ### Changed